+++ app/javascript/components/game/controls/buttons/DeleteGameButton.tsx
@@ -1,23 +1,31 @@
 import React, { FormEvent } from "react";
+import { XCircle } from "react-bootstrap-icons";
 import Game from "../../../../engine/Game";
-import { CancelGlyph } from "../../../utilities/buttons";
 import { OverlayTrigger, Tooltip, TooltipProps } from "react-bootstrap";
+import { useNavigate } from "react-router-dom";
 
-interface ResignCancelButtonProps {
+interface DeleteGameButtonProps {
   game: Game;
   vertical: boolean;
   callback: () => void;
 }
 
-export default function ResignCancelButton({ game, vertical, callback }: ResignCancelButtonProps) {
+export default function DeleteGameButton({ game, vertical, callback }: DeleteGameButtonProps) {
+  const navigate = useNavigate()
+
   const onSubmit = (event: FormEvent) => {
     event.preventDefault()
-    game.clearResignation()
-    game.clearDelete()
+    game.increaseDelete()
+    if (game.deleteLevel > 1) {
+      navigate("/", { replace: true })
+    }
     callback()
   }
 
-  const text = "cancel"
+  let text = "delete game"
+  if (vertical) {
+    text = game.deleteLevel > 0 ? "are you sure you want to delete this game?" : text
+  }
 
   const buttonTooltip = (props: TooltipProps) => (
     <Tooltip className="tooltip-game" {...props}>
@@ -32,11 +40,11 @@ export default function ResignCancelButton({ game, vertical, callback }: ResignC
           <OverlayTrigger placement="bottom" overlay={buttonTooltip}
                           delay={{ show: 0, hide: 0 }} >
             <button type="submit" className="custom-button custom-button-balance nowrap">
-              { CancelGlyph() }
+              <XCircle />
             </button>
           </OverlayTrigger> :
           <button type="submit" className="custom-button nowrap">
-            { CancelGlyph() } {text}
+            <XCircle /> {text}
           </button>
         }
       </div>